fix+feat: assignments_list float decode; modules_list with_items tree mode
Two issues from the user's earlier MCP test pass:

1. assignments_list(course_id) returned a decode error
     "invalid type: floating point `5.0`, expected i64 at line 1 column 4053"
   when no `bucket` filter was used. Canvas returns floats for the rubric
   `points` and ScoreStatistic stats fields; the Canvas-spec-driven generator
   typed them as i64. Patched the generated structs to f64 directly:
     - RubricCriteria.points        i64 -> f64
     - RubricCriterion.points       i64 -> f64
     - RubricRating.points          i64 -> f64
     - ScoreStatistic.{min,max,mean,upper_q,median,lower_q}  i64 -> f64
   The codegen template still emits i64 for these — the generator should be
   fixed in a follow-up; for now the manual patch is preserved with a
   comment so a regenerate doesn't silently undo it.

2. modules_list previously returned ModuleSummary with no items, so finding
   "which week contains HW3" required iterating modules_get per module.
   Added `with_items: bool` (default false): when true, calls Canvas's
   `?include[]=items` and returns [ModuleDetail] in one paginated round-trip.

Smoke tests:
  assignments_list(58272)               -> 4 assignments
  modules_list(58272, with_items=true)  -> 16 modules with items inline

+++ cool-tools/src/assignments.rs
@@ -118,7 +118,7 @@ pub async fn get_detail(
         .into_iter()
         .map(|c| ContractRubricCriterion {
             description: c.description.unwrap_or_default(),
-            points: c.points.unwrap_or(0) as f64,
+            points: c.points.unwrap_or(0.0),
             long_description: c.long_description,
         })
         .collect();
